T#772: Extract scheduler routes + auto-trigger daemon to src/scheduler/routes.ts
Phase 1.7 of Backend Modularization (Library #102). Mechanical move, no
logic changes.

**Moved to src/scheduler/routes.ts (~748L)**:
- parseInterval, parseDaysOfWeek (input validation)
- computeNextFixedTime, computeNextFixedTimeAfterRun, computeNextWeekdayFixedTime (UTC+7 anchored next-due computation)
- 4 singular /api/schedule routes (drizzle ORM via handleScheduleList/Add):
  GET /api/schedule/md, GET /api/schedule, POST /api/schedule, PATCH /api/schedule/:id
- 9 plural /api/schedules routes (sqlite-direct):
  GET /api/schedules, GET /api/schedules/due, GET /api/schedules/:id,
  POST /api/schedules, PATCH /api/schedules/:id, PATCH /api/schedules/:id/run,
  DELETE /api/schedules/:id, POST /api/schedules/:id/execute, PATCH /api/schedules/:id/trigger
- GET /api/scheduler/health
- runSchedulerCycle daemon function + SCHEDULER_INTERVAL constant + state
- Auto-trigger setInterval lifecycle (now inside initScheduler())
- Startup reset (triggered → pending) moved into initScheduler() — was top-level

**Exports**:
- registerSchedulerRoutes(app, sqlite, helpers) — DI: hasSessionAuth, requireBeastIdentity
- initScheduler(sqliteDb, drizzleDb, repoRoot, { wsBroadcast, enqueueNotification }) — captures deps, starts daemon (one-shot via schedulerStarted guard), runs startup-reset

**Server.ts changes**:
- Added import (L99) for both exports + initScheduler
- Removed singular /api/schedule block (L2929-2976, ~48L) — replaced with 1-line comment
- Removed plural /api/schedules + helpers + daemon (L6378-6995, ~618L) — replaced with 4 lines:
    initScheduler(sqlite, db, REPO_ROOT, { wsBroadcast, enqueueNotification });
    registerSchedulerRoutes(app, sqlite, { hasSessionAuth, requireBeastIdentity });

**Type safety**:
- Module-level state uses sentinel pattern: moduleSqlite (Database | null) shadowed by local
  const sqlite inside daemon body after null-guard; route handlers use const sqlite = sqliteDb
  closure to keep the verbatim plural block working unchanged.

**Verification**:
- Boot test: server starts clean, '[Scheduler] Auto-trigger daemon started (10s interval)'
  logs from new module, all endpoints listed including the 13 scheduler routes
- tsc --noEmit: zero new errors in src/scheduler/routes.ts (same OpenAPIHono<AppEnv>
  generic-param friction as 5/5 prior extractions — folded into T#788 sweep)
- No logic changes — only relocation. All auth gates preserved (PATCH /:id requester check,
  /run requireBeastIdentity, owner-or-Gorn ownership checks)
- Net: server.ts down ~660 lines (was 11963, now 10861 including merge with T#771)
